requirements.txt
streamlit==1.33.0
loguru==0.7.2
pdfplumber==0.11.0
rapidfuzz==3.7.0
python-dotenv==1.0.1

numpy==1.26.4

transformers==4.40.2
sentence-transformers==2.7.0

faiss-cpu==1.7.4

langchain==0.2.12
langchain-groq==0.1.4

tqdm==4.66.4
requests==2.32.3
